pyproject.toml
[project]
name = "tiny_python_projects"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "김백일", email = "cedar101@gmail.com"},
]
dependencies = [
    "docopt>=0.6.2",
    "python-box>=7.1.1",
    "pydantic>=2.6.1",
    "type-docopt>=0.8.2",
    "pronunciation-dictionary>=0.0.5",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black",
    "flake8",
    "ipython",
    "mypy",
    "pylint",
    "pytest",
    "yapf",
    "ptpython>=3.0.26",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
